+++ NeuralNets/Interfaces/ILossFunction.cs
@@ -48,7 +48,7 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
     // 1/2 * (precited - actual)^2
     public class SquaredLoss : ILossFunction
     {
-        public ColumnVector Error(ColumnVector truth, ColumnVector predicted) => 0.5 * (predicted - truth) * (predicted - truth);
+        public ColumnVector Error(ColumnVector truth, ColumnVector predicted) => 0.5F * (predicted - truth) * (predicted - truth);
 
         public ColumnVector Derivative(ColumnVector truth, ColumnVector predicted) => (predicted - truth);
     }
@@ -62,18 +62,18 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
 
         // not written for more than one sample.
         // Expects "predicted" to be the output from SoftMax
-        public double ScalarLoss(ColumnVector truth, ColumnVector predicted)
+        public float ScalarLoss(ColumnVector truth, ColumnVector predicted)
         {
             // use sum once, since this is one dimensional. For batch processing it'll be a 2d matrix and we'll use OneHotEconde (below)
             return this.Error(truth, predicted).Sum();
         }
 
-        public double ScalarLossBatch(Matrix truth, Matrix predicted)
+        public float ScalarLossBatch(Matrix truth, Matrix predicted)
         {
             Matrix oneHotSamples = this.OneHotEncode(truth);
             Matrix logYHAT = predicted.Log();
             Matrix YTimeYHat = oneHotSamples * logYHAT;
-            double sum = YTimeYHat.Sum();
+            float sum = YTimeYHat.Sum();
             return sum;
         }
 
@@ -82,7 +82,7 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
         public Matrix OneHotEncode(Matrix samples)
         {
             int N = samples.Rows;
-            double invN = -1.0 / (double)N;
+            float invN = -1 / (float)N;
 
             Matrix m = new Matrix(samples.Rows, samples.Cols);
             for (int i = 0; i < N; i++)
@@ -92,7 +92,7 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
                 {
                     if (samples[i, j] >= 0.99999)
                     {
-                        m[i, j] = 1.0;
+                        m[i, j] = 1;
                         break; // assumes there's only one. might want to assert this?  TODO
                     }
                 }
@@ -102,7 +102,7 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
 
         public ColumnVector Error(ColumnVector truth, ColumnVector predicted)
         {
-            return -1.0 * truth * predicted.Log();
+            return -1 * truth * predicted.Log();
         }
     }
 
@@ -131,9 +131,9 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
 
         // not written for more than one sample.
         // Expects "predicted" to be the output from SoftMax
-        public double ScalarLoss(ColumnVector truth, ColumnVector predicted)
+        public float ScalarLoss(ColumnVector truth, ColumnVector predicted)
         {
-            double NInv = -1.0;
+            float NInv = -1;
             ColumnVector logYHat = predicted.Log();
             return NInv * logYHat.Sum();
         }
@@ -141,15 +141,15 @@ Loss Function: Cross-Entropy, also referred to as Logarithmic loss.
         // not clear if the input needs to be an array with the actual value of the class, like [0,0,0,4,0,0,0,0] which means the 4th index is a four, that is we recognized a four in the image input.
         // totally broken, need to re-read this and get it right:
         // https://arjun-sarkar786.medium.com/implementation-of-all-loss-functions-deep-learning-in-numpy-tensorflow-and-pytorch-e20e72626ebd
-        public double ScalarLossBatch(Matrix truth, Matrix predicted)
+        public float ScalarLossBatch(Matrix truth, Matrix predicted)
         {
             throw new NotImplementedException();
             int N = truth.Rows;
-            double invN = -1.0 / (double)N;
+            float invN = -1 / (float)N;
 
             Matrix logYHAT = predicted.Log();
-            Matrix YTimeYHat = 1.0 * logYHAT;
-            double sum = YTimeYHat.Sum();
+            Matrix YTimeYHat = 1 * logYHAT;
+            float sum = YTimeYHat.Sum();
             return sum;
         }
